creat check package creat Check class update & debug code

+++ src/main/java/org/pharmacy/Main.java
@@ -1,5 +1,6 @@
 package org.pharmacy;
 
+import org.pharmacy.check.Check;
 import org.pharmacy.entities.Admin;
 import org.pharmacy.entities.Patient;
 import org.pharmacy.entities.Prescription;
@@ -15,93 +16,6 @@ import java.util.Scanner;
 
 public class Main {
     static Scanner input = new Scanner(System.in);
-
-    public static int check(int first, int last) {
-        int button;
-        while (true) {
-            if (input.hasNextInt()) {
-                int temp = input.nextInt();
-                if (temp >= first && temp <= last) {
-                    button = temp;
-                    return button;
-                } else {
-                    System.out.print("Enter Number Between" + first + "and" + last + " : ");
-                }
-            } else {
-                System.out.print("Enter Number! : ");
-                input.next();
-            }
-        }
-    }
-
-    public static int checkQuantity() {
-        int quantity;
-        while (true) {
-            if (input.hasNextInt()) {
-                int temp = input.nextInt();
-                if (temp > 0) {
-                    quantity = temp;
-                    return quantity;
-                } else {
-                    System.out.print("Enter Quantity Greater than 0! : ");
-                }
-            } else {
-                System.out.print("Enter Number! : ");
-                input.next();
-            }
-        }
-    }
-
-    public static String checkNationalCode() {
-        System.out.print("Enter national code : ");
-        String nationalCode;
-        while (true) {
-            if (input.hasNextInt()) {
-                String temp = input.next();
-                if (temp.length() == 10) {
-                    nationalCode = temp;
-                    return nationalCode;
-                } else {
-                    System.out.print("Enter 10 Number! : ");
-                }
-            } else {
-                System.out.print("Enter Number! : ");
-                input.next();
-            }
-        }
-    }
-    public static boolean checkYN(){
-        while (true){
-            String yn = input.next();
-            if (yn.equalsIgnoreCase("y") || yn.equalsIgnoreCase("n")){
-                if (yn.equalsIgnoreCase("y")){
-                    return true;
-                }else if (yn.equalsIgnoreCase("n")){
-                    return false;
-                }
-            }else {
-                System.out.println("Please Enter Y or N!");
-            }
-        }
-    }
-    public static int checkPrice() {
-        int price;
-        while (true) {
-            if (input.hasNextLong()) {
-                int temp = input.nextInt();
-                if (temp > 0) {
-                    price = temp;
-                    return price;
-                } else {
-                    System.out.print("Enter Price Greater than 0! : ");
-                }
-            } else {
-                System.out.print("Enter Number! : ");
-                input.next();
-            }
-        }
-    }
-
     public static void main(String[] args) throws SQLException {
         AdminService adminService = new AdminServiceImpl();
         PatientService patientService = new PatientServiceImpl();
@@ -113,7 +27,7 @@ public class Main {
         System.out.println("Admin : 2");
         System.out.println("Exit : 3");
         System.out.print("Enter Your Function : ");
-        int button = check(1, 3);
+        int button = Check.check(1,3,input);
         if (button == 1) {
             while (true) {
                 System.out.println("Patient Menu : ");
@@ -121,9 +35,9 @@ public class Main {
                 System.out.println("Sign In : 2");
                 System.out.println("Exit : 3");
                 System.out.print("Enter Your Function : ");
-                int button1 = check(1, 3);
+                int button1 = Check.check(1,3,input);
                 if (button1 == 1) {
-                    String nationalCode = checkNationalCode();
+                    String nationalCode = Check.checkNationalCode(input);
                     if (patientService.save(nationalCode)) {
                         System.out.println("Sing Up Successfully.");
                         patient = patientService.load(nationalCode);
@@ -134,7 +48,7 @@ public class Main {
                     }
                 }
                 if (button1 == 2) {
-                    String nationalCode = checkNationalCode();
+                    String nationalCode = Check.checkNationalCode(input);
                     if (patientService.load(nationalCode) != null) {
                         patient = patientService.load(nationalCode);
                         System.out.println("Welcome");
@@ -158,14 +72,14 @@ public class Main {
                     System.out.println("Show Prescription : 4");
                     System.out.println("Exit : 5");
                     System.out.print("Enter Your Function : ");
-                    int button2 = check(1, 5);
+                    int button2 = Check.check(1,5,input);
                     if (button2 == 1) {
                         Prescription prescription = prescriptionService.load(patient);
                         if (prescription == null || !prescription.getConfirm()) {
                             System.out.print("Enter Drug's Name : ");
                             String name = input.next();
                             System.out.print("Enter Drug,s Quantity : ");
-                            int quantity = checkQuantity();
+                            int quantity = Check.checkQuantity(input);
                             if (prescriptionService.save(name, quantity, patient)) {
                                 System.out.println("Drug is Add to Your Prescription");
                             } else {
@@ -177,7 +91,14 @@ public class Main {
                     }
                     if (button2 == 2) {
                         if (prescriptionService.load(patient) != null) {
-                            prescriptionService.remove(patient);
+                            System.out.print("Are You Sure to Remove Your Prescription ? (Y/N) : ");
+                            boolean yn =Check.checkYN(input);
+                            if (yn) {
+                                prescriptionService.remove(patient);
+                                System.out.println("Your Prescription Removed.");
+                            }else {
+                                System.out.println("Your Prescription Not Removed");
+                            }
                         }else {
                             System.out.println("Nothing to Remove!");
                         }
@@ -191,7 +112,7 @@ public class Main {
                                 System.out.print("Enter New Drug's Name : ");
                                 String newName = input.next();
                                 System.out.print("Enter New Drug,s Quantity : ");
-                                int newQuantity = checkQuantity();
+                                int newQuantity = Check.checkQuantity(input);
                                 prescriptionService.editPrescription(name, newName, newQuantity, patient);
                             }else {
                                 System.out.println("This Name is Not Exist In Your Prescription!");
@@ -226,7 +147,7 @@ public class Main {
                 System.out.println("Sign In : 2");
                 System.out.println("Exit : 3");
                 System.out.print("Enter Your Function : ");
-                int button3 = check(1, 3);
+                int button3 = Check.check(1,3,input);
                 if (button3 == 1) {
                     System.out.print("Enter username : ");
                     String username = input.next();
@@ -268,7 +189,7 @@ public class Main {
                     System.out.println("Accept Patient Payment : 3");
                     System.out.println("Exit : 4");
                     System.out.print("Enter Your Function : ");
-                    int button4 = check(1,4);
+                    int button4 = Check.check(1,4,input);
                     if (button4 == 1) {
                         Prescription prescription = prescriptionService.loadAll();
                         if (prescription != null){
@@ -278,44 +199,48 @@ public class Main {
                         }
                     }
                     if (button4 == 2) {
-                        String nationalCode = checkNationalCode();
-                        if (patientService.load(nationalCode) != null) {
-                            System.out.print("Are You Sure to Confirm " + nationalCode + " ?(Y/N)");
-                            boolean yn = checkYN();
-                            prescriptionService.changeMode("confirm", yn, nationalCode);
-                            if (yn){
-                                System.out.println(nationalCode + " Is Confirmed.");
-                            patient = patientService.load(nationalCode);
-                            Prescription prescription = prescriptionService.load(patient);
-                            String drugName;
-                            long price;
-                            if (prescription.size() > 0) {
-                                for (int i = 0; i < prescription.size(); i++) {
-                                    drugName = prescription.getName(i);
-                                    System.out.print("Does Exist " + drugName + " ?(Y/N)");
-                                    yn = checkYN();
-                                    prescriptionService.changeMode("exist", yn, drugName);
-                                    if (yn) {
-                                        System.out.print(drugName + " Price Is : ");
-                                        price = checkPrice();
-                                        prescriptionService.addPrice(drugName, price, nationalCode);
+                        String nationalCode = Check.checkNationalCode(input);
+                        patient = patientService.load(nationalCode);
+                        if (patient != null) {
+                            if (!prescriptionService.load(patient).getConfirm()) {
+                                System.out.print("Are You Sure to Confirm " + nationalCode + " ?(Y/N): ");
+                                boolean yn = Check.checkYN(input);
+                                prescriptionService.changeMode("confirm", yn, nationalCode);
+                                if (yn) {
+                                    System.out.println(nationalCode + " Is Confirmed.");
+                                    Prescription prescription = prescriptionService.load(patient);
+                                    String drugName;
+                                    long price;
+                                    if (prescription.size() > 0) {
+                                        for (int i = 0; i < prescription.size(); i++) {
+                                            drugName = prescription.getName(i);
+                                            System.out.print("Does Exist " + drugName + " ?(Y/N): ");
+                                            yn = Check.checkYN(input);
+                                            prescriptionService.changeMode("exist", yn, drugName);
+                                            if (yn) {
+                                                System.out.print(drugName + " Price Is : ");
+                                                price = Check.checkPrice(input);
+                                                prescriptionService.addPrice(drugName, price, nationalCode);
+                                            }
+                                        }
+                                    } else {
+                                        System.out.println("This Patient Don't Have any Prescription!");
                                     }
+                                } else {
+                                    System.out.println(nationalCode + " Is Not Confirmed.");
                                 }
-                            } else {
-                                System.out.println("This Patient Don't Have any Prescription!");
-                            }
-                        }else {
-                                System.out.println(nationalCode + " Is Not Confirmed.");
+                            }else {
+                                System.out.println(nationalCode + " Is Confirmed later");
                             }
                         }else {
                             System.out.println("This Patient Is Not Exist!");
                         }
                     }
                     if (button4 == 3){
-                        String nationalCode = checkNationalCode();
+                        String nationalCode = Check.checkNationalCode(input);
                         if (patientService.load(nationalCode) != null) {
-                            System.out.print("Are You Sure to Accept Payment for " + nationalCode + " ?(Y/N)");
-                            boolean yn = checkYN();
+                            System.out.print("Are You Sure to Accept Payment for " + nationalCode + " ?(Y/N): ");
+                            boolean yn = Check.checkYN(input);
                             prescriptionService.changeMode("pay", yn, nationalCode);
                             if (yn) {
                                 System.out.println(nationalCode + " Payment Accepted.");
